patch of autorun

+++ main.py
@@ -7,12 +7,13 @@ USER_NAME = getpass.getuser()
 
 
 def add_to_startup(file_path=""):
-    if file_path == "":
-        file_path = os.path.dirname(os.path.realpath(__file__))
-        file_path += "\output\main\main.exe"
+    file_path = os.path.dirname(os.path.realpath(__file__))
+    file_path += "\output\main\main.exe"
     bat_path = r'C:\Users\%s\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\Startup' % USER_NAME
-    with open(bat_path + '\\' + "open.bat", "w+") as bat_file:
-        bat_file.write(r'start %s' % file_path)
+    bp = bat_path + "\open.bat"
+    if os.path.exists(bp):
+        with open(bat_path + '\\' + "open.bat", "w+") as bat_file:
+            bat_file.write(r'start %s' % file_path)
 
 add_to_startup()
